--- portfolio/src/App.css
@@ -1,8 +1,0 @@
-.App {
-    border: 2px solid black;
-    text-align: center;
-    background: #f5f5f5;
-    color: #333;
-    margin: 20px;
-    padding: 20px;
-  }
