+++ portfolio/src/Components/App/App.css
@@ -1,0 +1,3 @@
+.container {
+  height: 100%;
+}
