+++ portfolio/src/Components/HomePage/HomePage.jsx
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+import "./HomePage.css";
+
+class HomePage extends Component {
+  render() {
+    return (
+      <div className = "homepage">
+        <h1 className = "homepage-heading">Navin Kumar</h1>
+        <p className = "homepage-content">Few Good lines about me. Think About it Later (:) .</p>
+      </div>
+    )
+  }
+}
+
+export default HomePage;
