+++ portfolio/src/index.jsx
@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
-import App from './Components/App';
+import App from './Components/App/App.jsx';
+
 ReactDOM.render(<App />, document.getElementById('app'));
